+++ tests/routes/test_helpers.py
@@ -1,0 +1,13 @@
+def test_viirs_vector_tile_server(client):
+    """
+    Basic test to check if empty data api response as expected
+    """
+
+    response = client.get("/_latest")
+    assert response.json() == {
+        "data": [
+            {"dataset": "nasa_viirs_fire_alerts", "version": "v202003"},
+            {"dataset": "wdpa_protected_areas", "version": "v201912"},
+        ],
+        "status": "success",
+    }
